requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
anthropic==0.18.1
openpyxl==3.1.2
python-docx==1.1.0
pandas==2.2.0
pdfplumber==0.10.3
